+++ packages/rocketh-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/web
 
+## 0.17.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.17.16
+  - rocketh@0.17.22
+
 ## 0.17.21
 
 ### Patch Changes